Draft: split cut from Draft.py
