+++ src/Mod/Draft/CMakeLists.txt
@@ -60,6 +60,7 @@ SET(Draft_utilities
 
 SET(Draft_functions
     draftfunctions/__init__.py
+    draftfunctions/cut.py
     draftfunctions/draftify.py
     draftfunctions/fuse.py
     draftfunctions/heal.py
