+++ src/Mod/Draft/draftfunctions/cut.py
@@ -1,0 +1,51 @@
+# ***************************************************************************
+# *   Copyright (c) 2009, 2010 Yorik van Havre <yorik@uncreated.net>        *
+# *   Copyright (c) 2009, 2010 Ken Cline <cline@frii.com>                   *
+# *   Copyright (c) 2020 FreeCAD Developers                                 *
+# *                                                                         *
+# *   This program is free software; you can redistribute it and/or modify  *
+# *   it under the terms of the GNU Lesser General Public License (LGPL)    *
+# *   as published by the Free Software Foundation; either version 2 of     *
+# *   the License, or (at your option) any later version.                   *
+# *   for detail see the LICENCE text file.                                 *
+# *                                                                         *
+# *   This program is distributed in the hope that it will be useful,       *
+# *   but WITHOUT ANY WARRANTY; without even the implied warranty of        *
+# *   MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the         *
+# *   GNU Library General Public License for more details.                  *
+# *                                                                         *
+# *   You should have received a copy of the GNU Library General Public     *
+# *   License along with this program; if not, write to the Free Software   *
+# *   Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  *
+# *   USA                                                                   *
+# *                                                                         *
+# ***************************************************************************
+"""This module provides the code for Draft cut function.
+"""
+## @package cut
+# \ingroup DRAFT
+# \brief This module provides the code for Draft cut function.
+
+import FreeCAD as App
+
+import draftutils.gui_utils as gui_utils
+
+
+def cut(object1,object2):
+    """cut(oject1,object2)
+    
+    Returns a cut object made from the difference of the 2 given objects.
+    """
+    if not App.ActiveDocument:
+        App.Console.PrintError("No active document. Aborting\n")
+        return
+    obj = App.ActiveDocument.addObject("Part::Cut","Cut")
+    obj.Base = object1
+    obj.Tool = object2
+    object1.ViewObject.Visibility = False
+    object2.ViewObject.Visibility = False
+    if App.GuiUp:
+        gui_utils.format_object(obj, object1)
+        gui_utils.select(obj)
+
+    return obj
